+++ cmd/pint/main.go
@@ -17,6 +17,8 @@ const (
 	disabledFlag = "disabled"
 	offlineFlag  = "offline"
 	noColorFlag  = "no-color"
+	intervalFlag = "interval"
+	listenFlag   = "listen"
 )
 
 var (
@@ -82,6 +84,37 @@ func newApp() *cli.App {
 						Aliases: []string{"o"},
 						Value:   false,
 						Usage:   "Disable all check that send live queries to Prometheus servers",
+					},
+				},
+			},
+			{
+				Name:   "watch",
+				Usage:  "Continuously lint specified files",
+				Action: actionWatch,
+				Flags: []cli.Flag{
+					&cli.StringSliceFlag{
+						Name:    disabledFlag,
+						Aliases: []string{"d"},
+						Value:   cli.NewStringSlice(),
+						Usage:   "List of checks to disable (example: promql/cost)",
+					},
+					&cli.BoolFlag{
+						Name:    offlineFlag,
+						Aliases: []string{"o"},
+						Value:   false,
+						Usage:   "Disable all check that send live queries to Prometheus servers",
+					},
+					&cli.DurationFlag{
+						Name:    intervalFlag,
+						Aliases: []string{"i"},
+						Value:   time.Minute * 10,
+						Usage:   "How often to run all checks",
+					},
+					&cli.StringFlag{
+						Name:    listenFlag,
+						Aliases: []string{"s"},
+						Value:   ":8080",
+						Usage:   "Listen address for HTTP web server exposing metrics",
 					},
 				},
 			},