+++ internal/reporter/github.go
@@ -7,10 +7,11 @@ import (
 
 	"context"
 
-	"github.com/cloudflare/pint/internal/git"
 	"github.com/google/go-github/v37/github"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/oauth2"
+
+	"github.com/cloudflare/pint/internal/git"
 )
 
 type GithubReporter struct {